feat: Implement SqlSchemaValidator for SQL query validation and add tests for various scenarios
